refactor #5 moved active_trail_nodes function
Fixes #5
